requirements.txt
torch
transformers
sentencepiece
numpy
pillow